Fix date dimension.
The date dimension was originally defined on a per-day basis, using
d3.time.day.  This resulted in date.top(40) returning the top 40 flights
when ordered by day, but within a day, the flights were not necessarily
ordered by time.

The fix is to define the date dimension using the full date and time,
and then define the group using d3.time.day instead.

Fixes #77, #24.